+++ android/TodokTodok/app/src/main/java/com/team/todoktodok/presentation/view/discussiondetail/commentcreate/CommentCreateBottomSheet.kt
@@ -1,6 +1,5 @@
-package com.team.todoktodok.presentation.view.discussiondetail
+package com.team.todoktodok.presentation.view.discussiondetail.commentcreate
 
-import android.app.Dialog
 import android.content.Context
 import android.content.DialogInterface
 import android.os.Bundle
@@ -14,12 +13,15 @@ import com.google.android.material.bottomsheet.BottomSheetDialogFragment
 import com.team.todoktodok.App
 import com.team.todoktodok.R
 import com.team.todoktodok.databinding.FragmentCommentCreateBottomSheetBinding
-import com.team.todoktodok.presentation.view.discussiondetail.vm.CommentCreateViewModel
+import com.team.todoktodok.presentation.view.discussiondetail.BottomSheetVisibilityListener
+import com.team.todoktodok.presentation.view.discussiondetail.CommentCreateUiEvent
+import com.team.todoktodok.presentation.view.discussiondetail.commentcreate.vm.CommentCreateViewModel
+import com.team.todoktodok.presentation.view.discussiondetail.commentcreate.vm.CommentCreateViewModelFactory
 
 class CommentCreateBottomSheet : BottomSheetDialogFragment(R.layout.fragment_comment_create_bottom_sheet) {
     private val viewModel by viewModels<CommentCreateViewModel> {
         val repositoryModule = (requireActivity().application as App).container.repositoryModule
-        com.team.todoktodok.presentation.view.discussiondetail.vm.CommentCreateViewModelFactory(
+        CommentCreateViewModelFactory(
             repositoryModule.commentRepository,
         )
     }
@@ -29,11 +31,6 @@ class CommentCreateBottomSheet : BottomSheetDialogFragment(R.layout.fragment_com
     fun setVisibilityListener(listener: BottomSheetVisibilityListener) {
         visibilityListener = listener
     }
-
-    override fun onCreateDialog(savedInstanceState: Bundle?): Dialog =
-        super.onCreateDialog(savedInstanceState).apply {
-            window?.setDimAmount(0f)
-        }
 
     override fun onStart() {
         super.onStart()
@@ -57,6 +54,8 @@ class CommentCreateBottomSheet : BottomSheetDialogFragment(R.layout.fragment_com
         visibilityListener?.onBottomSheetDismissed()
     }
 
+    override fun getTheme(): Int = R.style.BottomSheetDialogTheme
+
     private fun initView(binding: FragmentCommentCreateBottomSheetBinding) {
         with(binding) {
             etTextCommentContent.requestFocus()
@@ -77,10 +76,10 @@ class CommentCreateBottomSheet : BottomSheetDialogFragment(R.layout.fragment_com
     }
 
     private fun setupObserve(binding: FragmentCommentCreateBottomSheetBinding) {
-        viewModel.uiEvent.observe(this) { value ->
+        viewModel.uiEvent.observe(viewLifecycleOwner) { value ->
             handleUiEvent(value)
         }
-        viewModel.commentText.observe(this) { value ->
+        viewModel.commentText.observe(viewLifecycleOwner) { value ->
             binding.ivAddComment.isEnabled = value.isNotBlank()
         }
     }
@@ -105,7 +104,7 @@ class CommentCreateBottomSheet : BottomSheetDialogFragment(R.layout.fragment_com
     companion object {
         const val COMMENT_REQUEST_KEY = "comment_create_result"
         const val COMMENT_CREATED_RESULT_KEY = "IS_COMMENT_CREATED"
-        const val TAG = "COMMENT_CREATE_BOTTOM_SHEET"
+        const val TAG = "COMMENTS_BOTTOM_SHEET"
 
         fun newInstance(discussionId: Long): CommentCreateBottomSheet =
             CommentCreateBottomSheet().apply {
