+++ android/TodokTodok/app/src/main/java/com/team/todoktodok/presentation/view/discussiondetail/commentcreate/vm/CommentCreateViewModel.kt
@@ -1,4 +1,4 @@
-package com.team.todoktodok.presentation.view.discussiondetail.vm
+package com.team.todoktodok.presentation.view.discussiondetail.commentcreate.vm
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
